Move emitter methods to the files they are used in
